refactor: extract useWorkspaceWindowActions hook from App.tsx
